move resizing function to hook
