+++ client/landing/stepper/declarative-flow/internals/steps-repository/site-migration-instructions/site-preview/index.tsx
@@ -1,6 +1,6 @@
 import { SiteThumbnail, Spinner } from '@automattic/components';
 import { useTranslate } from 'i18n-calypso';
-import { FC, useEffect, useRef } from 'react';
+import React, { FC } from 'react';
 import { useQuery } from 'calypso/landing/stepper/hooks/use-query';
 import { useSitePreviewMShotImageHandler } from './hooks/use-site-preview-mshot-image-handler';
 import './style.scss';
@@ -9,22 +9,7 @@ export const SitePreview: FC = () => {
 	const translate = useTranslate();
 	const fromUrl = useQuery().get( 'from' ) || '';
 
-	const previewRef = useRef< HTMLDivElement >( null );
-	const { mShotsOption, updateDimensions } = useSitePreviewMShotImageHandler();
-
-	useEffect( () => {
-		updateDimensions( previewRef );
-
-		// Throttle the resize event handling.
-		let resizeTimeout: ReturnType< typeof setTimeout >;
-		const throttledResizeHandler = () => {
-			clearTimeout( resizeTimeout );
-			resizeTimeout = setTimeout( () => updateDimensions( previewRef ), 200 );
-		};
-
-		window.addEventListener( 'resize', throttledResizeHandler );
-		return () => window.removeEventListener( 'resize', throttledResizeHandler );
-	}, [] );
+	const { mShotsOption, previewRef } = useSitePreviewMShotImageHandler();
 
 	return (
 		<div className="migration-instructions-from-preview" ref={ previewRef }>
